+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/room/OutboxMessageEventHandler.kt
@@ -21,7 +21,6 @@ import de.connect2x.trixnity.client.utils.retryLoop
 import de.connect2x.trixnity.clientserverapi.client.MatrixClientServerApiClient
 import de.connect2x.trixnity.clientserverapi.model.media.FileTransferProgress
 import de.connect2x.trixnity.core.EventHandler
-import de.connect2x.trixnity.core.MSC4354
 import de.connect2x.trixnity.core.MatrixServerException
 import de.connect2x.trixnity.core.UserInfo
 import de.connect2x.trixnity.core.model.RoomId
@@ -29,7 +28,7 @@ import de.connect2x.trixnity.core.model.events.MessageEventContent
 import de.connect2x.trixnity.core.model.events.m.MarkedUnreadEventContent
 import de.connect2x.trixnity.core.subscribe
 import de.connect2x.trixnity.core.unsubscribeOnCompletion
-import io.ktor.http.HttpStatusCode
+import io.ktor.http.*
 import kotlin.time.Clock
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.CoroutineScope
@@ -162,7 +161,6 @@ class OutboxMessageEventHandler(
         log.debug { "cancel sending of ${outboxMessage.transactionId}" }
     }
 
-    @OptIn(MSC4354::class)
     private suspend fun sendOutboxMessage(outboxMessage: RoomOutboxMessage<*>, roomId: RoomId): SendError? {
         val transactionId = outboxMessage.transactionId
         log.trace { "send outbox message (transactionId=${transactionId}, roomId=${outboxMessage.roomId})" }
@@ -247,12 +245,7 @@ class OutboxMessageEventHandler(
             try {
                 log.debug { "send outbox message $transactionId into $roomId" }
                 api.room
-                    .sendMessageEvent(
-                        roomId = roomId,
-                        eventContent = content,
-                        txnId = transactionId,
-                        stickyDurationMs = outboxMessage.stickyDuration?.inWholeMilliseconds,
-                    )
+                    .sendMessageEvent(roomId = roomId, eventContent = content, txnId = transactionId)
                     .getOrThrow() // TODO fold as soon as continue is supported in inline lambdas
             } catch (exception: MatrixServerException) {
                 val sendError =